Main code refactor, fix some ui issues, code overview and clean
