--- src/main/java/pl/medos/cmmsApi/exception/SensorNotFoundException.java
@@ -1,5 +1,0 @@
-package pl.medos.cmmsApi.exception;
-
-public class SensorNotFoundException extends Exception{
-    public SensorNotFoundException(String message){super(message);}
-}
