+++ src/main/java/pl/medos/cmmsApi/model/Part.java
@@ -1,55 +1,17 @@
 package pl.medos.cmmsApi.model;
 
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
 public class Part {
 
     private Long id;
     private String name;
     private int quantity;
-
-//    private Supplier supplier;
-//    private Material materials;
-//    private Process processes;
-//    private Cost cost;
-
-    public Part() {
-    }
-
-    public Part(Long id, String name, int quantity) {
-        this.id = id;
-        this.name = name;
-        this.quantity = quantity;
-    }
-
-    public Long getId() {
-        return id;
-    }
-
-    public void setId(Long id) {
-        this.id = id;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public int getQuantity() {
-        return quantity;
-    }
-
-    public void setQuantity(int quantity) {
-        this.quantity = quantity;
-    }
-
-    @Override
-    public String toString() {
-        return "Part{" +
-                "id=" + id +
-                ", name='" + name + '\'' +
-                ", quantity=" + quantity +
-                '}';
-    }
 }
