+++ src/main/java/pl/medos/cmmsApi/model/Stock.java
@@ -1,5 +1,14 @@
 package pl.medos.cmmsApi.model;
 
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
 public class Stock {
 
     private Long id;
@@ -7,7 +16,5 @@ public class Stock {
     private int stockRegistry;
     private Part part;
     private Cost cost;
-    public Stock() {
-    }
 
 }
