+++ src/main/java/pl/medos/cmmsApi/repository/entity/DepartmentEntity.java
@@ -1,7 +1,17 @@
 package pl.medos.cmmsApi.repository.entity;
 
 import jakarta.persistence.*;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
 
+@Getter
+@Setter
+@NoArgsConstructor
+@AllArgsConstructor
+@Builder
 @Entity
 @Table(name = "DEPARTMENTS")
 public class DepartmentEntity {
@@ -12,45 +22,4 @@ public class DepartmentEntity {
     private String name;
     private String location;
 
-    public DepartmentEntity() {
-    }
-
-    public DepartmentEntity(Long id, String name, String location) {
-        this.id = id;
-        this.name = name;
-        this.location = location;
-    }
-
-    public Long getId() {
-        return id;
-    }
-
-    public void setId(Long id) {
-        this.id = id;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public String getLocation() {
-        return location;
-    }
-
-    public void setLocation(String location) {
-        this.location = location;
-    }
-
-    @Override
-    public String toString() {
-        return "DepartmentEntity{" +
-                "id=" + id +
-                ", name='" + name + '\'' +
-                ", location='" + location + '\'' +
-                '}';
-    }
 }
