+++ src/main/java/pl/medos/cmmsApi/repository/entity/InvoiceEntity.java
@@ -1,8 +1,17 @@
 package pl.medos.cmmsApi.repository.entity;
 
 import jakarta.persistence.*;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
 import pl.medos.cmmsApi.model.Supplier;
-
+@Getter
+@Setter
+@NoArgsConstructor
+@AllArgsConstructor
+@Builder
 @Entity
 @Table(name = "INVOICES")
 public class InvoiceEntity {
@@ -17,48 +26,4 @@ public class InvoiceEntity {
     @JoinColumn(name = "supplier_id")
     private SupplierEntity supplier;
 
-    public InvoiceEntity() {
-    }
-
-    public Long getId() {
-        return id;
-    }
-
-    public void setId(Long id) {
-        this.id = id;
-    }
-
-    public String getNumber() {
-        return number;
-    }
-
-    public void setNumber(String number) {
-        this.number = number;
-    }
-
-    public double getPrice() {
-        return price;
-    }
-
-    public void setPrice(double price) {
-        this.price = price;
-    }
-
-    public SupplierEntity getSupplier() {
-        return supplier;
-    }
-
-    public void setSupplier(SupplierEntity supplier) {
-        this.supplier = supplier;
-    }
-
-    @Override
-    public String toString() {
-        return "InvoiceEntity{" +
-                "id=" + id +
-                ", number='" + number + '\'' +
-                ", price=" + price +
-                ", supplier=" + supplier +
-                '}';
-    }
 }
