+++ src/main/java/pl/medos/cmmsApi/repository/entity/PartEntity.java
@@ -1,7 +1,17 @@
 package pl.medos.cmmsApi.repository.entity;
 
 import jakarta.persistence.*;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
 
+@Getter
+@Setter
+@NoArgsConstructor
+@AllArgsConstructor
+@Builder
 @Entity
 @Table(name = "PARTS")
 public class PartEntity {
@@ -11,50 +21,4 @@ public class PartEntity {
     private String name;
     private int quantity;
 
-//    private Supplier supplier;
-//    private Material materials;
-//    private Process processes;
-//    private Cost cost;
-
-    public PartEntity() {
-    }
-
-    public PartEntity(Long id, String name, int quantity) {
-        this.id = id;
-        this.name = name;
-        this.quantity = quantity;
-    }
-
-    public Long getId() {
-        return id;
-    }
-
-    public void setId(Long id) {
-        this.id = id;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public int getQuantity() {
-        return quantity;
-    }
-
-    public void setQuantity(int quantity) {
-        this.quantity = quantity;
-    }
-
-    @Override
-    public String toString() {
-        return "Part{" +
-                "id=" + id +
-                ", name='" + name + '\'' +
-                ", quantity=" + quantity +
-                '}';
-    }
 }
