+++ src/main/java/pl/medos/cmmsApi/repository/entity/ResourceEntity.java
@@ -1,7 +1,17 @@
 package pl.medos.cmmsApi.repository.entity;
 
 import jakarta.persistence.*;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
 
+@Getter
+@Setter
+@NoArgsConstructor
+@AllArgsConstructor
+@Builder
 @Entity
 @Table(name = "RESOURCES")
 public class ResourceEntity {
@@ -16,48 +26,4 @@ public class ResourceEntity {
     @JoinColumn(name = "invoice_id")
     private InvoiceEntity invoice;
 
-    public ResourceEntity() {
-    }
-
-    public InvoiceEntity getInvoice() {
-        return invoice;
-    }
-
-    public void setInvoice(InvoiceEntity invoice) {
-        this.invoice = invoice;
-    }
-
-    public Long getId() {
-        return id;
-    }
-
-    public void setId(Long id) {
-        this.id = id;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public double getQuantity() {
-        return quantity;
-    }
-
-    public void setQuantity(double quantity) {
-        this.quantity = quantity;
-    }
-
-    @Override
-    public String toString() {
-        return "ResourceEntity{" +
-                "id=" + id +
-                ", name='" + name + '\'' +
-                ", quantity=" + quantity +
-                ", invoice=" + invoice +
-                '}';
-    }
 }
