+++ src/main/resources/templates/create-job.html
@@ -31,7 +31,7 @@
                                        class="block text-xl font-medium text-red-600 sm:mt-px sm:pt-2">Informacje o
                                     zgłoszeniu</label>
                                 <div class="mt-1 sm:col-span-2 sm:mt-0">
-                                    <textarea id="message" name="message" rows="3"
+                                    <textarea th:field="*{message}" id="message" name="message" rows="3"
                                               class="block w-full max-w-lg rounded-md border-gray-300 shadow-sm focus:border-indigo-500 focus:ring-indigo-500 sm:text-sm"></textarea>
                                     <p class="mt-2 text-sm text-gray-700">Wprowadź informację, czego dotyczy awaria lub
                                         zlecenie.</p>
