+++ src/main/resources/templates/list-machine.html
@@ -1,6 +1,6 @@
 <!DOCTYPE html >
 <html lang="en" xmlns="http://www.w3.org/1999/xhtml"
-      xmlns:th="http://www.thymeleaf.org"
+      xmlns:th="http://www.thymeleaf.org">
 <head>
     <meta charset="UTF-8">
     <title>Moduł Pracowników</title>
@@ -372,8 +372,8 @@
                                             <tbody class="bg-grey">
                                             <tr class="border-t border-gray-200">
                                                 <th colspan="8" scope="colgroup"
-                                                    class="bg-grey-700 px-4 py-2 text-left text-lm font-semibold text-gray-900 sm:px-6">
-                                                    Maszyny wycofane z eksploatacji
+                                                    class="bg-grey-700 px-4 py-2 text-center text-lm font-semibold text-gray-500 sm:px-6">
+                                                    Koniec zestawienia
                                                 </th>
                                             </tr>
                                         </table>
